allow autocomplete to be used without a map instance
This enables the geocoder autocomplete to be used in a standalone
manner, without a map instance tied to it.

Note that although it is now possible to use the autocomplete
without a map instance, the Mapbox Terms of Service requires that
POI search results still be used with a Mapbox map.

Per the Mapbox geocoder docs:

> Keep in mind that the Mapbox Terms of Service require that POI
> search results be shown on a Mapbox map. If you don't need POIs,
> you can exclude them from your search results with the
> options.types parameter when constructing a new Geocoder.

See https://github.com/mapbox/mapbox-gl-geocoder/tree/e02134da433f435aa827f16321a327483fd7217a#using-without-a-map